Fazendo apis de Carrinho, Pratos, Extras...

+++ backend/app/models/bebida_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column,Boolean, Integer, String, ForeignKey, DateTime
+from sqlalchemy import Column,Boolean,Numeric, Integer, String, ForeignKey, DateTime
 from datetime import datetime
 import uuid
 from sqlalchemy.orm import relationship
@@ -9,14 +9,14 @@ class Bebida(Base):
 
     id = Column(String(36), primary_key=True, default=lambda: str(uuid.uuid4()))
     nome = Column(String(100), nullable=False)
-    preco = Column(Integer, nullable=False)
-    precoDesconto = Column(Integer, nullable=True)
+    preco = Column(Numeric(10, 2), nullable=False)
+    precoDesconto = Column(Numeric(10, 2), nullable=True)
     desconto = Column(Boolean, default=False)
     descricao = Column(String(255), nullable=False)
     imagem_url = Column(String(255), nullable=False)
     estrela = Column(Integer, nullable=True)
     legenda = Column(String(100), nullable=False)
-    unidadeMedida = Column(Integer, nullable=False)
+    unidadeMedida = Column(Numeric(10, 1), nullable=False)
 
     criado_em = Column(DateTime, default=datetime.utcnow)
     atualizado_em = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
