+++ backend/app/models/pedido_models.py
@@ -1,6 +1,6 @@
 import uuid
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime
+from sqlalchemy import Column, Integer,Numeric , String, ForeignKey, DateTime
 from sqlalchemy.orm import relationship
 from app.database import Base
 
@@ -11,10 +11,10 @@ class Pedido(Base):
     item_id = Column(Integer, ForeignKey("itens.id"), nullable=False)
     usuario_id = Column(String(36), ForeignKey("usuarios.id"), nullable=False)
     endereco_id = Column(Integer, ForeignKey("enderecos.id"), nullable=False)
-    frete = Column(Integer, nullable=True)
+    frete = Column(Numeric(10, 2), nullable=True)
     retirada_loja = Column(Integer, nullable=True, default=0)
     retirada_hora = Column(String(50), nullable=True)
-    total = Column(Integer, nullable=False)
+    total = Column(Numeric(10, 2), nullable=False)
     status = Column(String(50), nullable=False, default="Pendente")
     observacoes = Column(String(255), nullable=True)
